Comprobación de permisos para items / refactorización codigo

+++ src/controllers/items.controller.js
@@ -1,4 +1,7 @@
 const Item = require('../models/Items')
+const Lista = require('../models/Listas')
+const Usuario = require('../models/Usuarios')
+const jwt = require('jsonwebtoken')
 
  const obtenerItems = async (req, res) => {
     try {
@@ -10,8 +13,9 @@ const Item = require('../models/Items')
  }
 
  const obtenerItem = async (req, res) => {
+    const { id } = req.params
+    
     try {
-        const { id } = req.params
         const item = await Item.findOne({
             where: {id}
         })
@@ -22,8 +26,30 @@ const Item = require('../models/Items')
  }
 
  const crearItem = async (req, res) => {
+    const { nombre, cantidadConseguida, cantidadNecesitada, listaId } = req.body
+    const { token } = req.cookies
+
+    let id
     try {
-        const { nombre, cantidadConseguida, cantidadNecesitada, listaId } = req.body
+        id = jwt.verify(token, process.env.JWT_SECRET).id
+    } catch (error) {
+        res.status(401).json({ message: 'Token inválido o expirado' });
+    }
+
+    try {
+        const lista = await Usuario.findOne({
+            where: { id },
+            include: {
+                model: Lista,
+                where: { id: listaId }
+            },
+            attributes: []
+        })
+
+        if (!lista || !lista.Listas.length === 0) {
+            return res.sendStatus(404)
+        }
+
         const item = await Item.create({
             nombre,
             cantidadConseguida,
@@ -37,26 +63,84 @@ const Item = require('../models/Items')
  }
 
  const actualizarItem = async (req, res) => {
+    const { token } = req.cookies
+    const { id } = req.params
+    const { nombre, cantidadNecesitada, cantidadConseguida } = req.body
+
+    let info
     try {
-        const { id } = req.params
-        const { nombre, cantidadNecesitada, cantidadConseguida } = req.body
-        const item = await Item.update({
+        info = jwt.verify(token, process.env.JWT_SECRET)
+    } catch (error) {
+        return res.status(401).json({ message: 'Token inválido o expirado' });
+    }
+
+    try {
+        const item = await Item.findOne({
+            where: { id }
+        })
+        
+        if (!item) {
+            return res.sendStatus(404)
+        }
+
+        const usuarioConItemEnLista = await Usuario.findOne({
+            where: { id: info.id },
+            include: {
+                model: Lista,
+                where: { id: item.listaId }
+            }
+        })
+
+        if (!usuarioConItemEnLista) {
+            return res.sendStatus(404)
+        }
+
+        const itemActualizado = await Item.update({
             nombre,
             cantidadNecesitada,
             cantidadConseguida
         }, {
             where: { id }
         })
-        res.json(item)
+        res.json(itemActualizado)
     } catch (error) {
         res.sendStatus(500)
     }
  }
 
  const eliminarItem = async (req, res) => {
+    const { token } = req.cookies
+    const { id } = req.params
+
+    let info;
     try {
-        const { id } = req.params
-        const item = await Item.destroy({
+        info = jwt.verify(token, process.env.JWT_SECRET)
+    } catch (error) {
+        return res.status(401).json({ message: 'Token inválido o expirado' });
+    }
+
+    try {
+        const item = await Item.findOne({
+            where: { id }
+        })
+        
+        if (!item) {
+            return res.sendStatus(404)
+        }
+
+        const usuarioConItemEnLista = await Usuario.findOne({
+            where: { id: info.id },
+            include: {
+                model: Lista,
+                where: { id: item.listaId }
+            }
+        })
+
+        if (!usuarioConItemEnLista) {
+            return res.sendStatus(404)
+        }
+
+        await Item.destroy({
             where: { id }
         })
         res.sendStatus(204)
